- make Posterior the data unit (delete ToCudaDataLoaders)
- add imputation method (+ update basic tutorial)

- corrupt on the fly for imputation benchmark
