+++ scvi/benchmark.py
@@ -2,13 +2,12 @@ import numpy as np
 from sklearn.decomposition import PCA
 
 from scvi.dataset import CortexDataset
-from scvi.inference import UnsupervisedTrainer, TrainerFish, adversarial_wrapper, ImputationTrainer
-from scvi.inference.classifier_inference import compute_accuracy_nn
-from scvi.inference.posterior import proximity_imputation
+from scvi.inference import UnsupervisedTrainer, TrainerFish, adversarial_wrapper
+from scvi.inference.posterior import proximity_imputation, compute_accuracy_nn
 from scvi.models import VAE, VAEF
 
 
-def cortex_benchmark(n_epochs=250, use_cuda=True, unit_test=False):
+def cortex_benchmark(n_epochs=250, use_cuda=True):
     cortex_dataset = CortexDataset()
     vae = VAE(cortex_dataset.nb_genes)
     infer_cortex_vae = UnsupervisedTrainer(vae, cortex_dataset, use_cuda=use_cuda)
@@ -16,10 +15,15 @@ def cortex_benchmark(n_epochs=250, use_cuda=True, unit_test=False):
 
     infer_cortex_vae.test_set.ll()  # assert ~ 1200
     infer_cortex_vae.test_set.differential_expression()
-    imputation_trainer = ImputationTrainer(infer_cortex_vae, cortex_dataset, rate=0.1)
-    imputation_trainer.train(n_epochs)
-    n_samples = 1000 if not unit_test else 10
-    infer_cortex_vae.sequential.show_t_sne(n_samples=n_samples)
+    vae = VAE(cortex_dataset.nb_genes)
+    infer_cortex_vae = UnsupervisedTrainer(vae, cortex_dataset, use_cuda=use_cuda)
+    infer_cortex_vae.corrupt_posteriors()
+    infer_cortex_vae.train(n_epochs=n_epochs)
+    infer_cortex_vae.uncorrupt_posteriors()
+    infer_cortex_vae.train_set.imputation_benchmark(verbose=(n_epochs > 1))
+
+    n_samples = 10 if n_epochs == 1 else None  # n_epochs == 1 is unit tests
+    infer_cortex_vae.train_set.show_t_sne(n_samples=n_samples)
     return infer_cortex_vae
