+++ scvi/inference/__init__.py
@@ -1,28 +1,25 @@
 from .trainer import Trainer
-from .trainers import (
+from .inference import (
     UnsupervisedTrainer,
     TrainerFish,
-    ImputationTrainer,
     AdapterTrainer
 )
 from .posterior import Posterior
-from .classifier_inference import (
+from .annotation import (
     JointSemiSupervisedTrainer,
     SemiSupervisedTrainer,
     AlternateSemiSupervisedTrainer,
     ClassifierTrainer
 )
-from .experimental_inference import adversarial_wrapper, mmd_wrapper
+from .experimental_inference import adversarial_wrapper
 
 
 __all__ = ['UnsupervisedTrainer',
            'Trainer',
            'Posterior',
            'TrainerFish',
-           'ImputationTrainer',
            'AdapterTrainer',
            'adversarial_wrapper',
-           'mmd_wrapper',
            'JointSemiSupervisedTrainer',
            'SemiSupervisedTrainer',
            'AlternateSemiSupervisedTrainer',
