+++ scvi/inference/posterior.py
@@ -1,8 +1,12 @@
+import copy
+from collections import namedtuple
+
 import numpy as np
 import scipy
 import torch
 from matplotlib import pyplot as plt
 from scipy.stats import kde, entropy, itemfreq
+from sklearn import neighbors
 from sklearn.cluster import KMeans
 from sklearn.manifold import TSNE
 from sklearn.metrics import adjusted_rand_score as ARI
@@ -11,46 +15,201 @@ from sklearn.metrics import silhouette_score
 from sklearn.mixture import GaussianMixture as GMM
 from sklearn.neighbors import NearestNeighbors, KNeighborsRegressor
 from sklearn.utils.linear_assignment_ import linear_assignment
+from torch.utils.data import DataLoader
+from torch.utils.data.sampler import SequentialSampler, SubsetRandomSampler, RandomSampler
 
 from scvi.dataset import CortexDataset
 from scvi.models.log_likelihood import compute_log_likelihood, compute_marginal_log_likelihood
 
 
-def unsupervised_clustering_accuracy(y, y_pred):
-    """
-    Unsupervised Clustering Accuracy
-    """
-    assert len(y_pred) == len(y)
-    n_clusters = len(np.unique(y))
-    reward_matrix = np.zeros((n_clusters, n_clusters), dtype=np.int64)
-    for y_pred_, y_ in zip(y_pred, y):
-        reward_matrix[y_pred_, y_] += 1
-    cost_matrix = reward_matrix.max() - reward_matrix
-    ind = linear_assignment(cost_matrix)
-    return sum([reward_matrix[i, j] for i, j in ind]) * 1.0 / y_pred.size, ind
+class SequentialSubsetSampler(SubsetRandomSampler):
+    def __iter__(self):
+        return iter(self.indices)
 
 
 class Posterior:
-    def __init__(self, model, gene_dataset, data_loader, name='', to_monitor=[]):
+    def __init__(self, model, gene_dataset, shuffle=False, indices=None, use_cuda=True, data_loader_kwargs=dict()):
+        '''
+
+        When added to annotation, has a private name attribute
+        '''
         self.model = model
         self.gene_dataset = gene_dataset
-        self.data_loader = data_loader
-        self.name = name
-        self.to_monitor = to_monitor
+        self.to_monitor = []
+        self.use_cuda = use_cuda
+
+        if indices is not None and shuffle:
+            raise ValueError('indices is mutually exclusive with shuffle')
+        if indices is None:
+            if shuffle:
+                sampler = RandomSampler(gene_dataset)
+            else:
+                sampler = SequentialSampler(gene_dataset)
+        else:
+            sampler = SubsetRandomSampler(indices)
+        self.data_loader_kwargs = copy.copy(data_loader_kwargs)
+        self.data_loader_kwargs.update({'collate_fn': gene_dataset.collate_fn, 'sampler': sampler})
+        self.data_loader = DataLoader(gene_dataset, **self.data_loader_kwargs)
+
+    @property
+    def indices(self):
+        if hasattr(self.data_loader.sampler, 'indices'):
+            return self.data_loader.sampler.indices
+        else:
+            return np.arange(len(self.gene_dataset))
+
+    def __iter__(self):
+        return map(self.to_cuda, iter(self.data_loader))
+
+    def to_cuda(self, tensors):
+        return [t.cuda(async=self.use_cuda) if self.use_cuda else t for t in tensors]
+
+    def update(self, data_loader_kwargs):
+        posterior = copy.copy(self)
+        posterior.data_loader_kwargs = copy.copy(self.data_loader_kwargs)
+        posterior.data_loader_kwargs.update(data_loader_kwargs)
+        posterior.data_loader = DataLoader(self.gene_dataset, **posterior.data_loader_kwargs)
+        return posterior
+
+    def sequential(self, batch_size=128):
+        return self.update({'batch_size': batch_size, 'sampler': SequentialSubsetSampler(indices=self.indices)})
+
+    def corrupted(self):
+        return self.update({'collate_fn': self.gene_dataset.collate_fn_corrupted})
+
+    def uncorrupted(self):
+        return self.update({'collate_fn': self.gene_dataset.collate_fn})
 
     def ll(self, verbose=False):
-        ll = compute_log_likelihood(self.model, self.data_loader)
+        ll = compute_log_likelihood(self.model, self)
         if verbose:
-            print("LL %s is : %.4f" % (self.name, ll))
+            print("LL : %.4f" % ll)
         return ll
 
     ll.mode = 'min'
 
     def marginal_ll(self, verbose=False, n_mc_samples=1000):
-        ll = compute_marginal_log_likelihood(self.model, self.data_loader, n_mc_samples)
+        ll = compute_marginal_log_likelihood(self.model, self, n_mc_samples)
         if verbose:
-            print("True LL for %s is : %.4f" % (self.name, ll))
+            print("True LL : %.4f" % ll)
         return ll
+
+    def get_latent(self):
+        latent = []
+        batch_indices = []
+        labels = []
+        for tensors in self:
+            sample_batch, local_l_mean, local_l_var, batch_index, label = tensors
+            latent += [self.model.sample_from_posterior_z(sample_batch, y=label)]
+            batch_indices += [batch_index]
+            labels += [label]
+        return np.array(torch.cat(latent)), np.array(torch.cat(batch_indices)), np.array(torch.cat(labels)).ravel()
+
+    def entropy_batch_mixing(self, verbose=False, **kwargs):
+        if self.gene_dataset.n_batches == 2:
+            latent, batch_indices, labels = self.get_latent()
+            be_score = entropy_batch_mixing(latent, batch_indices, **kwargs)
+            if verbose:
+                print("Entropy batch mixing :", be_score)
+            return be_score
+
+    entropy_batch_mixing.mode = 'max'
+
+    def differential_expression(self, *args, verbose=False, **kwargs):
+        px_scale, all_labels = self.differential_expression_stats(*args, **kwargs)
+
+        if type(self.gene_dataset) == CortexDataset:
+            if 'use_cuda' in kwargs:
+                kwargs.pop('use_cuda')
+            de_score = de_cortex(px_scale, all_labels, self.gene_dataset.gene_names, **kwargs)
+            if verbose:
+                print("DE score for cortex: %.4f" % de_score)
+            return de_score
+
+    differential_expression.mode = 'max'
+
+    def differential_expression_stats(self, M_sampling=100):
+        """
+        Output average over statistics in a symmetric way (a against b)
+        forget the sets if permutation is True
+        :param vae: The generative vae and encoder network
+        :param data_loader: a data loader for a particular dataset
+        :param M_sampling: number of samples
+        :return: A 1-d vector of statistics of size n_genes
+        """
+        px_scales = []
+        all_labels = []
+        for tensors in self:
+            sample_batch, _, _, batch_index, labels = tensors
+            sample_batch = sample_batch.repeat(1, M_sampling).view(-1, sample_batch.size(1))
+            batch_index = batch_index.repeat(1, M_sampling).view(-1, 1)
+            labels = labels.repeat(1, M_sampling).view(-1, 1)
+            px_scales += [
+                (self.model.get_sample_scale(sample_batch, batch_index=batch_index, y=labels).squeeze()).cpu()]
+            all_labels += [labels.cpu()]
+
+        px_scale = torch.cat(px_scales)
+        all_labels = torch.cat(all_labels)
+
+        return px_scale, all_labels
+
+    def imputation(self, n_samples=1):
+        imputed_list = []
+        for tensors in self:
+            sample_batch, _, _, batch_index, labels = tensors
+            px_rate = self.model.get_sample_rate(sample_batch, batch_index=batch_index, y=labels, n_samples=n_samples)
+            imputed_list += [np.array(px_rate.cpu())]
+        imputed_list = np.concatenate(imputed_list, axis=1)
+        return imputed_list.squeeze()
+
+    def imputation_benchmark(self, n_samples=8, verbose=False):
+        original_list = []
+        imputed_list = []
+        batch_size = self.data_loader_kwargs['batch_size'] // n_samples
+        for tensors, corrupted_tensors in zip(self.sequential(batch_size=batch_size),
+                                              self.corrupted().sequential(batch_size=batch_size)):
+            batch = tensors[0]
+            actual_batch_size = batch.size(0)
+            dropout_batch, _, _, batch_index, labels = corrupted_tensors
+            px_rate = self.model.get_sample_rate(dropout_batch, batch_index=batch_index, y=labels, n_samples=n_samples)
+
+            indices_dropout = torch.nonzero(batch - dropout_batch)
+            i = indices_dropout[:, 0]
+            j = indices_dropout[:, 1]
+
+            batch = batch.unsqueeze(0).expand((n_samples, batch.size(0), batch.size(1)))
+            original = np.array(batch[:, i, j].view(-1).cpu())
+            imputed = np.array(px_rate[:, i, j].view(-1).cpu())
+
+            cells_index = np.tile(np.array(i.cpu()), n_samples)
+
+            original_list += [original[cells_index == i] for i in range(actual_batch_size)]
+            imputed_list += [imputed[cells_index == i] for i in range(actual_batch_size)]
+
+        # Median of medians for all distances
+        median_score = np.median(np.abs(np.concatenate(original_list) - np.concatenate(imputed_list)))
+
+        # Mean of medians for each cell
+        imputation_cells = []
+        for original, imputed in zip(original_list, imputed_list):
+            has_imputation = len(original) and len(imputed)
+            imputation_cells += [np.median(np.abs(original - imputed)) if has_imputation else 0]
+        mean_score = np.mean(imputation_cells)
+
+        if verbose:
+            print("\nMedian of Median: %.4f\nMean of Median for each cell: %.4f" % (median_score, mean_score))
+
+        plot_imputation(np.concatenate(original_list), np.concatenate(imputed_list))
+        return original_list, imputed_list
+
+    def knn_purity(self, verbose=False):
+        latent, _, labels = self.get_latent()
+        score = knn_purity(latent, labels)
+        if verbose:
+            print("KNN purity score :", score)
+        return score
+
+    knn_purity.mode = 'max'
 
     def clustering_scores(self, verbose=True, prediction_algorithm='knn'):
         if self.gene_dataset.n_labels > 1:
@@ -67,88 +226,65 @@ class Posterior:
             ari_score = ARI(labels, labels_pred)
             uca_score = unsupervised_clustering_accuracy(labels, labels_pred)[0]
             if verbose:
-                print("Clustering Scores for %s:\nSilhouette: %.4f\nNMI: %.4f\nARI: %.4f\nUCA: %.4f" %
-                      (self.name, asw_score, nmi_score, ari_score, uca_score))
+                print("Clustering Scores:\nSilhouette: %.4f\nNMI: %.4f\nARI: %.4f\nUCA: %.4f" %
+                      (asw_score, nmi_score, ari_score, uca_score))
             return asw_score, nmi_score, ari_score, uca_score
 
     def nn_overlap_score(self, verbose=True, **kwargs):
         if hasattr(self.gene_dataset, 'adt_expression_clr'):
-            assert self.name == 'sequential'  # only works for the sequential data_loader (mapping indices)
-            latent, _, _ = self.get_latent()
-            protein_data = self.gene_dataset.adt_expression_clr
+            latent, _, _ = self.sequential().get_latent()
+            protein_data = self.gene_dataset.adt_expression_clr[self.indices]
             spearman_correlation, fold_enrichment = nn_overlap(latent, protein_data, **kwargs)
             if verbose:
-                print("Overlap Scores for %s:\nSpearman Correlation: %.4f\nFold Enrichment: %.4f" %
-                      (self.name, spearman_correlation, fold_enrichment))
+                print("Overlap Scores:\nSpearman Correlation: %.4f\nFold Enrichment: %.4f" %
+                      (spearman_correlation, fold_enrichment))
             return spearman_correlation, fold_enrichment
 
-    def differential_expression(self, *args, verbose=False, **kwargs):
-        px_scale, all_labels = self.de_stats(*args, **kwargs)
-
-        if type(self.gene_dataset) == CortexDataset:
-            if 'use_cuda' in kwargs:
-                kwargs.pop('use_cuda')
-            de_score = de_cortex(px_scale, all_labels, self.gene_dataset.gene_names, **kwargs)
-            if verbose:
-                print("DE score for cortex on %s is : %.4f" % (self.name, de_score))
-            return de_score
-
-    differential_expression.mode = 'max'
-
-    def entropy_batch_mixing(self, verbose=False, **kwargs):
-        if self.gene_dataset.n_batches == 2:
-            latent, batch_indices, labels = self.get_latent()
-            be_score = entropy_batch_mixing(latent, batch_indices, **kwargs)
-            if verbose:
-                print("Entropy batch mixing :", be_score)
-            return be_score
-
-    entropy_batch_mixing.mode = 'max'
-
-    def knn_purity(self, verbose=False):
-        latent, _, labels = self.get_latent()
-        score = knn_purity(latent, labels)
+    def accuracy(self, verbose=False):
+        model, cls = (self.sampling_model, self.model) if hasattr(self, 'sampling_model') else (self.model, None)
+        acc = compute_accuracy(model, self, classifier=cls)
         if verbose:
-            print("KNN purity score :", score)
-        return score
+            print("Acc: %.4f" % (acc))
+        return acc
 
-    knn_purity.mode = 'max'
+    accuracy.mode = 'max'
 
-    def get_latent(self):
-        latent = []
-        batch_indices = []
-        labels = []
-        for tensors in self.data_loader:
-            sample_batch, local_l_mean, local_l_var, batch_index, label = tensors
-            latent += [self.model.sample_from_posterior_z(sample_batch, y=label)]
-            batch_indices += [batch_index]
-            labels += [label]
-        return np.array(torch.cat(latent)), np.array(torch.cat(batch_indices)), np.array(torch.cat(labels)).ravel()
+    def hierarchical_accuracy(self, name, verbose=False):
 
-    def de_stats(self, M_sampling=100):
-        """
-        Output average over statistics in a symmetric way (a against b)
-        forget the sets if permutation is True
-        :param vae: The generative vae and encoder network
-        :param data_loader: a data loader for a particular dataset
-        :param M_sampling: number of samples
-        :return: A 1-d vector of statistics of size n_genes
-        """
-        px_scales = []
-        all_labels = []
-        for tensors in self.data_loader:
-            sample_batch, _, _, batch_index, labels = tensors
-            sample_batch = sample_batch.repeat(1, M_sampling).view(-1, sample_batch.size(1))
-            batch_index = batch_index.repeat(1, M_sampling).view(-1, 1)
-            labels = labels.repeat(1, M_sampling).view(-1, 1)
-            px_scales += [
-                (self.model.get_sample_scale(sample_batch, batch_index=batch_index, y=labels).squeeze()).cpu()]
-            all_labels += [labels.cpu()]
+        all_y, all_y_pred = compute_predictions(self.model, self)
+        acc = np.mean(all_y == all_y_pred)
 
-        px_scale = torch.cat(px_scales)
-        all_labels = torch.cat(all_labels)
+        all_y_groups = np.array([self.model.labels_groups[y] for y in all_y])
+        all_y_pred_groups = np.array([self.model.labels_groups[y] for y in all_y_pred])
+        h_acc = np.mean(all_y_groups == all_y_pred_groups)
 
-        return px_scale, all_labels
+        if verbose:
+            print("Acc for %s is : %.4f\nHierarchical Acc for %s is : %.4f\n" % (name, acc, name, h_acc))
+        return acc
+
+    accuracy.mode = 'max'
+
+    def unsupervised_accuracy(self, verbose=False):
+        uca = unsupervised_classification_accuracy(self.model, self)[0]
+        if verbose:
+            print("UCA : %.4f" % (uca))
+        return uca
+
+    unsupervised_accuracy.mode = 'max'
+
+    def ll_fish(self, verbose=False):
+        ll = compute_log_likelihood(self.model, self, mode="smFISH")
+        if verbose:
+            print("LL Fish: %.4f" % ll)
+        return ll
+
+    def show_spatial_expression(self, x_coord, y_coord, labels, color_by='scalar', title='spatial_expression.svg'):
+        x_coord = x_coord.reshape(-1, 1)
+        y_coord = y_coord.reshape(-1, 1)
+        latent = np.concatenate((x_coord, y_coord), axis=1)
+        self.show_t_sne(name=None, n_samples=1000, color_by=color_by, save_name=title,
+                        latent=latent, batch_indices=None,
+                        labels=labels)
 
     def show_t_sne(self, n_samples=1000, color_by='', save_name='', latent=None, batch_indices=None,
                    labels=None, n_batch=None):
@@ -208,91 +344,23 @@ class Posterior:
             latent = TSNE().fit_transform(latent[idx_t_sne])
         return latent, idx_t_sne
 
-    @staticmethod
-    def raw_data(*posteriors):
+    def raw_data(self):
         """
-        From a list of data_loaders, return the numpy array suitable for classification with sklearn
-        :param data_loaders: a sequence of data_loaders arguments
-        :return: a sequence of (data, labels) for each data_loader
+        Returns raw data for classification
         """
-
-        def get_data_labels(data_loader):
-            if hasattr(data_loader, 'sampler') and hasattr(data_loader.sampler, 'indices'):
-                data = data_loader.dataset.X[data_loader.sampler.indices]
-                labels = data_loader.dataset.labels[data_loader.sampler.indices]
-            else:
-                data = data_loader.dataset.X
-                labels = data_loader.dataset.labels
-            return data, labels.ravel()
-
-        return [get_data_labels(posterior.data_loader) for posterior in posteriors]
+        return self.gene_dataset.X[self.indices], self.gene_dataset.labels[self.indices].ravel()
 
 
-class PosteriorFish(Posterior):
-    def ll(self, verbose=False):
-        ll = compute_log_likelihood(self.model, self.data_loader, mode="smFISH")
-        if verbose:
-            print("LL for %s is : %.4f" % (self.name, ll))
-        return ll
-
-    def show_spatial_expression(self, x_coord, y_coord, labels, color_by='scalar', title='spatial_expression.svg'):
-        x_coord = x_coord.reshape(-1, 1)
-        y_coord = y_coord.reshape(-1, 1)
-        latent = np.concatenate((x_coord, y_coord), axis=1)
-        self.show_t_sne(name=None, n_samples=1000, color_by=color_by, save_name=title,
-                        latent=latent, batch_indices=None,
-                        labels=labels)
+def entropy_from_indices(indices):
+    return entropy(np.array(itemfreq(indices)[:, 1].astype(np.int32)))
 
 
-def plot_imputation(original, imputed, title="Imputation"):
-    y = imputed
-    x = original
+def entropy_batch_mixing(latent_space, batches, n_neighbors=50, n_pools=50, n_samples_per_pool=100, max_number=500):
+    # max number is an important parameter
+    n_samples = len(latent_space)
+    keep_idx = np.random.choice(np.arange(n_samples), size=min(len(latent_space), max_number), replace=False)
+    latent_space, batches = latent_space[keep_idx], batches[keep_idx]
 
-    ymax = 10
-    mask = x < ymax
-    x = x[mask]
-    y = y[mask]
-
-    mask = y < ymax
-    x = x[mask]
-    y = y[mask]
-
-    l_minimum = np.minimum(x.shape[0], y.shape[0])
-
-    x = x[:l_minimum]
-    y = y[:l_minimum]
-
-    data = np.vstack([x, y])
-
-    plt.figure(figsize=(5, 5))
-
-    axes = plt.gca()
-    axes.set_xlim([0, ymax])
-    axes.set_ylim([0, ymax])
-
-    nbins = 50
-
-    # Evaluate a gaussian kde on a regular grid of nbins x nbins over data extents
-    k = kde.gaussian_kde(data)
-    xi, yi = np.mgrid[0:ymax:nbins * 1j, 0:ymax:nbins * 1j]
-    zi = k(np.vstack([xi.flatten(), yi.flatten()]))
-
-    plt.title(title, fontsize=12)
-    plt.ylabel("Imputed counts")
-    plt.xlabel('Original counts')
-
-    plt.pcolormesh(yi, xi, zi.reshape(xi.shape), cmap="Reds")
-
-    a, _, _, _ = np.linalg.lstsq(y[:, np.newaxis], x, rcond=-1)
-    linspace = np.linspace(0, ymax)
-    plt.plot(linspace, a * linspace, color='black')
-
-    plt.plot(linspace, linspace, color='black', linestyle=":")
-    plt.show()
-    plt.savefig(title + '.png')
-
-
-def entropy_batch_mixing(latent_space, batches, n_neighbors=50, n_pools=50, n_samples_per_pool=100):
     batches = batches.ravel()
     nbrs = NearestNeighbors(n_neighbors=n_neighbors + 1).fit(latent_space)
     indices = nbrs.kneighbors(latent_space, return_distance=False)[:, 1:]
@@ -309,40 +377,6 @@ def entropy_batch_mixing(latent_space, batches, n_neighbors=50, n_pools=50, n_sa
         ])
 
     return score
-
-
-def entropy_from_indices(indices):
-    return entropy(np.array(itemfreq(indices)[:, 1].astype(np.int32)))
-
-
-def knn_purity(latent, label, n_neighbors=30):
-    nbrs = NearestNeighbors(n_neighbors=n_neighbors + 1).fit(latent)
-    indices = nbrs.kneighbors(latent, return_distance=False)[:, 1:]
-    neighbors_labels = np.vectorize(lambda i: label[i])(indices)
-
-    # pre cell purity scores
-    scores = ((neighbors_labels - label.reshape(-1, 1)) == 0).mean(axis=1)
-    res = [np.mean(scores[label == i]) for i in np.unique(label)]  # per cell-type purity
-
-    return np.mean(res)
-
-
-def nn_overlap(X1, X2, k=100):
-    nne = NearestNeighbors(n_neighbors=k + 1, n_jobs=8)
-    assert len(X1) == len(X2)
-    n_samples = len(X1)
-    nne.fit(X1)
-    kmatrix_1 = nne.kneighbors_graph(X1) - scipy.sparse.identity(n_samples)
-    nne.fit(X2)
-    kmatrix_2 = nne.kneighbors_graph(X2) - scipy.sparse.identity(n_samples)
-
-    # 1 - spearman correlation from knn graphs
-    spearman_correlation = scipy.stats.spearmanr(kmatrix_1.A.flatten(), kmatrix_2.A.flatten())[0]
-    # 2 - fold enrichment
-    set_1 = set(np.where(kmatrix_1.A.flatten() == 1)[0])
-    set_2 = set(np.where(kmatrix_2.A.flatten() == 1)[0])
-    fold_enrichment = len(set_1.intersection(set_2)) * n_samples ** 2 / (float(len(set_1)) * len(set_2))
-    return spearman_correlation, fold_enrichment
 
 
 def de_cortex(px_scale, all_labels, gene_names, M_permutation=100000, permutation=False):
@@ -393,6 +427,169 @@ def de_cortex(px_scale, all_labels, gene_names, M_permutation=100000, permutatio
     result = [(gene_name, res[np.where(gene_names == gene_name)[0]][0]) for gene_name in genes_of_interest]
     print('\n'.join([gene_name + " : " + str(r) for (gene_name, r) in result]))
     return result[1][1]  # if we had to give a metric to optimize
+
+
+def plot_imputation(original, imputed, title="Imputation"):
+    y = imputed
+    x = original
+
+    ymax = 10
+    mask = x < ymax
+    x = x[mask]
+    y = y[mask]
+
+    mask = y < ymax
+    x = x[mask]
+    y = y[mask]
+
+    l_minimum = np.minimum(x.shape[0], y.shape[0])
+
+    x = x[:l_minimum]
+    y = y[:l_minimum]
+
+    data = np.vstack([x, y])
+
+    plt.figure(figsize=(5, 5))
+
+    axes = plt.gca()
+    axes.set_xlim([0, ymax])
+    axes.set_ylim([0, ymax])
+
+    nbins = 50
+
+    # Evaluate a gaussian kde on a regular grid of nbins x nbins over data extents
+    k = kde.gaussian_kde(data)
+    xi, yi = np.mgrid[0:ymax:nbins * 1j, 0:ymax:nbins * 1j]
+    zi = k(np.vstack([xi.flatten(), yi.flatten()]))
+
+    plt.title(title, fontsize=12)
+    plt.ylabel("Imputed counts")
+    plt.xlabel('Original counts')
+
+    plt.pcolormesh(yi, xi, zi.reshape(xi.shape), cmap="Reds")
+
+    a, _, _, _ = np.linalg.lstsq(y[:, np.newaxis], x, rcond=-1)
+    linspace = np.linspace(0, ymax)
+    plt.plot(linspace, a * linspace, color='black')
+
+    plt.plot(linspace, linspace, color='black', linestyle=":")
+    plt.show()
+    plt.savefig(title + '.png')
+
+
+def nn_overlap(X1, X2, k=100):
+    nne = NearestNeighbors(n_neighbors=k + 1, n_jobs=8)
+    assert len(X1) == len(X2)
+    n_samples = len(X1)
+    nne.fit(X1)
+    kmatrix_1 = nne.kneighbors_graph(X1) - scipy.sparse.identity(n_samples)
+    nne.fit(X2)
+    kmatrix_2 = nne.kneighbors_graph(X2) - scipy.sparse.identity(n_samples)
+
+    # 1 - spearman correlation from knn graphs
+    spearman_correlation = scipy.stats.spearmanr(kmatrix_1.A.flatten(), kmatrix_2.A.flatten())[0]
+    # 2 - fold enrichment
+    set_1 = set(np.where(kmatrix_1.A.flatten() == 1)[0])
+    set_2 = set(np.where(kmatrix_2.A.flatten() == 1)[0])
+    fold_enrichment = len(set_1.intersection(set_2)) * n_samples ** 2 / (float(len(set_1)) * len(set_2))
+    return spearman_correlation, fold_enrichment
+
+
+def unsupervised_clustering_accuracy(y, y_pred):
+    """
+    Unsupervised Clustering Accuracy
+    """
+    assert len(y_pred) == len(y)
+    n_clusters = len(np.unique(y))
+    reward_matrix = np.zeros((n_clusters, n_clusters), dtype=np.int64)
+    for y_pred_, y_ in zip(y_pred, y):
+        reward_matrix[y_pred_, y_] += 1
+    cost_matrix = reward_matrix.max() - reward_matrix
+    ind = linear_assignment(cost_matrix)
+    return sum([reward_matrix[i, j] for i, j in ind]) * 1.0 / y_pred.size, ind
+
+
+def knn_purity(latent, label, n_neighbors=30):
+    nbrs = NearestNeighbors(n_neighbors=n_neighbors + 1).fit(latent)
+    indices = nbrs.kneighbors(latent, return_distance=False)[:, 1:]
+    neighbors_labels = np.vectorize(lambda i: label[i])(indices)
+
+    # pre cell purity scores
+    scores = ((neighbors_labels - label.reshape(-1, 1)) == 0).mean(axis=1)
+    res = [np.mean(scores[label == i]) for i in np.unique(label)]  # per cell-type purity
+
+    return np.mean(res)
+
+
+Accuracy = namedtuple('Accuracy', ['unweighted', 'weighted', 'worst', 'accuracy_classes'])
+
+
+def compute_accuracy_tuple(y, y_pred):
+    y = y.ravel()
+    n_labels = len(np.unique(y))
+    classes_probabilities = []
+    accuracy_classes = []
+    for cl in range(n_labels):
+        idx = y == cl
+        classes_probabilities += [np.mean(idx)]
+        accuracy_classes += [np.mean((y[idx] == y_pred[idx])) if classes_probabilities[-1] else 0]
+        # This is also referred to as the "recall": p = n_true_positive / (n_false_negative + n_true_positive)
+        # ( We could also compute the "precision": p = n_true_positive / (n_false_positive + n_true_positive) )
+        accuracy_named_tuple = Accuracy(
+            unweighted=np.dot(accuracy_classes, classes_probabilities),
+            weighted=np.mean(accuracy_classes),
+            worst=np.min(accuracy_classes),
+            accuracy_classes=accuracy_classes)
+    return accuracy_named_tuple
+
+
+def compute_predictions(model, data_loader, classifier=None):
+    all_y_pred = []
+    all_y = []
+
+    for i_batch, tensors in enumerate(data_loader):
+        sample_batch, _, _, _, labels = tensors
+        all_y += [labels.view(-1)]
+
+        if hasattr(model, 'classify'):
+            y_pred = model.classify(sample_batch).argmax(dim=-1)
+        elif classifier is not None:
+            # Then we use the specified classifier
+            if model is not None:
+                sample_batch, _, _ = model.z_encoder(sample_batch)
+            y_pred = classifier(sample_batch).argmax(dim=-1)
+        else:  # The model is the raw classifier
+            y_pred = model(sample_batch).argmax(dim=-1)
+        all_y_pred += [y_pred]
+
+    all_y_pred = np.array(torch.cat(all_y_pred))
+    all_y = np.array(torch.cat(all_y))
+    return all_y, all_y_pred
+
+
+def compute_accuracy(vae, data_loader, classifier=None):
+    all_y, all_y_pred = compute_predictions(vae, data_loader, classifier=classifier)
+    return np.mean(all_y == all_y_pred)
+
+
+def unsupervised_classification_accuracy(vae, data_loader, classifier=None):
+    all_y, all_y_pred = compute_predictions(vae, data_loader, classifier=classifier)
+    return unsupervised_clustering_accuracy(all_y, all_y_pred)
+
+
+def compute_accuracy_nn(data_train, labels_train, data_test, labels_test, k=5):
+    clf = neighbors.KNeighborsClassifier(k, weights='distance')
+    return compute_accuracy_classifier(clf, data_train, labels_train, data_test, labels_test)
+
+
+def compute_accuracy_classifier(clf, data_train, labels_train, data_test, labels_test):
+    clf.fit(data_train, labels_train)
+    # Predicting the labels
+    y_pred_test = clf.predict(data_test)
+    y_pred_train = clf.predict(data_train)
+
+    return (compute_accuracy_tuple(labels_train, y_pred_train),
+            compute_accuracy_tuple(labels_test, y_pred_test)), y_pred_test
 
 
 def proximity_imputation(real_latent1, normed_gene_exp_1, real_latent2, k=4):
