+++ scvi/models/__init__.py
@@ -3,8 +3,8 @@ from .vae import VAE
 from .vaec import VAEC
 from .vae_fish import VAEF
 from.classifier import Classifier
-__all__ = ['VAEC',
-           'SVAEC',
+__all__ = ['SVAEC',
+           'VAEC',
            'VAE',
            'VAEF',
            'Classifier']
